parent state refactroing
