+++ src/app/header/progress-bar/progress-bar.component.html
@@ -6,7 +6,8 @@
     isLoadingAdminData ||
     isLoadingMainPage ||
     isLoadingProvider ||
-    isAutorizationLoading
+    isAutorizationLoading ||
+    isLoadingParent
   "
   class="progress-bar">
   <mat-progress-bar mode="indeterminate"> </mat-progress-bar>
