+++ src/app/shell/details/details-tabs/reviews/reviews.component.ts
@@ -1,3 +1,4 @@
+import { ParentState } from './../../../../shared/store/parent.state.';
 import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { Actions, ofActionCompleted, Select, Store } from '@ngxs/store';
@@ -18,16 +19,15 @@ import { MetaDataState } from 'src/app/shared/store/meta-data.state';
 import { OnPageChangeRating, SetRatingPerPage } from 'src/app/shared/store/paginator.actions';
 import { PaginatorState } from 'src/app/shared/store/paginator.state';
 import { RegistrationState } from 'src/app/shared/store/registration.state';
-import { CreateRating, GetReviewedApplications, GetStatusAllowedToReview, OnCreateRatingSuccess } from 'src/app/shared/store/user.actions';
-import { UserState } from 'src/app/shared/store/user.state';
+import { CreateRating, GetReviewedApplications, GetStatusAllowedToReview, OnCreateRatingSuccess } from 'src/app/shared/store/parent.actions';
 
 @Component({
   selector: 'app-reviews',
   templateUrl: './reviews.component.html',
-  styleUrls: ['./reviews.component.scss']
+  styleUrls: ['./reviews.component.scss'],
 })
 export class ReviewsComponent implements OnInit, OnDestroy {
-  readonly noResultReviews= NoResultsTitle.noReviews;
+  readonly noResultReviews = NoResultsTitle.noReviews;
   readonly Role: typeof Role = Role;
   readonly ReviewDeclination = ReviewDeclination;
 
@@ -36,9 +36,9 @@ export class ReviewsComponent implements OnInit, OnDestroy {
 
   @Select(RegistrationState.parent)
   parent$: Observable<Parent>;
-  @Select(UserState.isAllowedToReview)
+  @Select(ParentState.isAllowedToReview)
   isAllowedToReview$: Observable<boolean>;
-  @Select(UserState.isReviewed)
+  @Select(ParentState.isReviewed)
   isReviewed$: Observable<boolean>;
 
   @Select(MetaDataState.rating)
@@ -55,54 +55,55 @@ export class ReviewsComponent implements OnInit, OnDestroy {
   isReviewed: boolean;
   currentPage: PaginationElement = PaginationConstants.firstPage;
 
-  constructor(
-    private store: Store,
-    private matDialog: MatDialog,
-    private actions$: Actions,
-  ) { }
+  constructor(private store: Store, private matDialog: MatDialog, private actions$: Actions) {}
 
   ngOnInit(): void {
     this.getParentData();
     this.getWorkshopRatingList();
 
-    this.actions$.pipe(ofActionCompleted(OnCreateRatingSuccess))
-      .pipe(
-        takeUntil(this.destroy$),
-        distinctUntilChanged())
-      .subscribe(() => this.store.dispatch([new GetRateByEntityId(EntityType.workshop, this.workshop.id), new GetReviewedApplications(this.parent.id, this.workshop.id)]));
+    this.actions$
+      .pipe(ofActionCompleted(OnCreateRatingSuccess))
+      .pipe(takeUntil(this.destroy$), distinctUntilChanged())
+      .subscribe(() =>
+        this.store.dispatch([
+          new GetRateByEntityId(EntityType.workshop, this.workshop.id),
+          new GetReviewedApplications(this.parent.id, this.workshop.id),
+        ])
+      );
 
     this.isAllowedToReview$
-    .pipe(takeUntil(this.destroy$))
-    .subscribe((status: boolean) => (this.isAllowedToReview = status));
+      .pipe(takeUntil(this.destroy$))
+      .subscribe((status: boolean) => (this.isAllowedToReview = status));
 
-    this.isReviewed$.pipe(
-      takeUntil(this.destroy$)
-    ).subscribe((status: boolean) => (this.isReviewed = status));
+    this.isReviewed$.pipe(takeUntil(this.destroy$)).subscribe((status: boolean) => (this.isReviewed = status));
 
-    this.ratingPerPage$.pipe(
-      takeUntil(this.destroy$)
-    ).subscribe((ratingPerPage: number) => (this.ratingPerPage = ratingPerPage));
+    this.ratingPerPage$
+      .pipe(takeUntil(this.destroy$))
+      .subscribe((ratingPerPage: number) => (this.ratingPerPage = ratingPerPage));
   }
 
   private getParentData(): void {
-    this.parent$.pipe(
-      filter((parent: Parent) => !!parent),
-      takeUntil(this.destroy$)
-    ).subscribe((parent: Parent) => {
-      this.parent = parent;
-      this.store.dispatch([
-        new GetStatusAllowedToReview(this.parent.id, this.workshop.id),
-        new GetReviewedApplications(this.parent.id, this.workshop.id)
-      ]);
-    });
+    this.parent$
+      .pipe(
+        filter((parent: Parent) => !!parent),
+        takeUntil(this.destroy$)
+      )
+      .subscribe((parent: Parent) => {
+        this.parent = parent;
+        this.store.dispatch([
+          new GetStatusAllowedToReview(this.parent.id, this.workshop.id),
+          new GetReviewedApplications(this.parent.id, this.workshop.id),
+        ]);
+      });
   }
 
   private getWorkshopRatingList(): void {
     this.rating$
       .pipe(
         filter((rating: Rate[]) => !!rating?.length),
-        takeUntil(this.destroy$),
-      ).subscribe((rating: Rate[]) => {
+        takeUntil(this.destroy$)
+      )
+      .subscribe((rating: Rate[]) => {
         this.rating = rating;
       });
   }
@@ -112,24 +113,29 @@ export class ReviewsComponent implements OnInit, OnDestroy {
       width: Constants.MODAL_SMALL,
       data: {
         type: ModalConfirmationType.rate,
-      }
+      },
     });
 
     dialogRef.afterClosed().subscribe((result: number) => {
       if (result) {
-        this.store.dispatch(new CreateRating({
-          rate: result,
-          type: Constants.WORKSHOP_ENTITY_TYPE,
-          entityId: `${this.workshop.id}`,
-          parentId: this.parent.id,
-        }));
+        this.store.dispatch(
+          new CreateRating({
+            rate: result,
+            type: Constants.WORKSHOP_ENTITY_TYPE,
+            entityId: `${this.workshop.id}`,
+            parentId: this.parent.id,
+          })
+        );
       }
     });
   }
-  
+
   itemsPerPageChange(itemsPerPage: number): void {
     this.ratingPerPage = itemsPerPage;
-    this.store.dispatch([new SetRatingPerPage(itemsPerPage), new GetRateByEntityId(EntityType.workshop, this.workshop.id)]);
+    this.store.dispatch([
+      new SetRatingPerPage(itemsPerPage),
+      new GetRateByEntityId(EntityType.workshop, this.workshop.id),
+    ]);
   }
 
   pageChange(page: PaginationElement): void {
