+++ src/app/shell/personal-cabinet/shared-cabinet/applications/applications.component.ts
@@ -9,13 +9,13 @@ import {
   ApplicationParameters,
   ApplicationUpdate,
 } from '../../../../shared/models/application.model';
-import { UpdateApplication } from 'src/app/shared/store/user.actions';
+import { UpdateApplication } from 'src/app/shared/store/shared-user.actions';
 import { MatTabChangeEvent } from '@angular/material/tabs/tab-group';
 import { MatTabGroup } from '@angular/material/tabs';
 import { NoResultsTitle } from 'src/app/shared/enum/no-results';
 import { ApplicationTitles, ApplicationTitlesReverse } from 'src/app/shared/enum/enumUA/applications';
 import { ActivatedRoute, Params, Router } from '@angular/router';
-import { OnUpdateApplicationSuccess } from '../../../../shared/store/user.actions';
+import { OnUpdateApplicationSuccess } from '../../../../shared/store/shared-user.actions';
 import { Observable, Subject } from 'rxjs';
 import { PaginatorState } from 'src/app/shared/store/paginator.state';
 import { PaginationElement } from 'src/app/shared/models/paginationElement.model';
@@ -23,7 +23,7 @@ import { OnPageChangeApplications, SetApplicationsPerPage } from 'src/app/shared
 import { PushNavPath } from 'src/app/shared/store/navigation.actions';
 import { NavBarName } from 'src/app/shared/enum/navigation-bar';
 import { ApplicationStatus } from 'src/app/shared/enum/applications';
-import { UserState } from 'src/app/shared/store/user.state';
+import { SharedUserState } from 'src/app/shared/store/shared-user.state';
 import { CabinetDataComponent } from '../cabinet-data.component';
 import { Child } from 'src/app/shared/models/child.model';
 import { Workshop } from 'src/app/shared/models/workshop.model';
@@ -41,12 +41,12 @@ export class ApplicationsComponent implements OnInit, OnDestroy, AfterViewInit {
   readonly noApplicationTitle = NoResultsTitle.noApplication;
   readonly Role = Role;
 
-  @Select(UserState.applications)
+  @Select(SharedUserState.applications)
   applicationCards$: Observable<ApplicationCards>;
   @Select(PaginatorState.applicationsPerPage)
   applicationsPerPage$: Observable<number>;
   applicationCards: ApplicationCards;
-  @Select(UserState.isLoading)
+  @Select(SharedUserState.isLoading)
   isLoadingCabinet$: Observable<boolean>;
 
   @ViewChild(MatTabGroup) tabGroup: MatTabGroup;
